+++ src/utils.rs
@@ -28,7 +28,7 @@ pub fn iter_neighbors(coord @ (x, y): Coordinate, height: usize, width: usize) -
 
 
 /// Returns a string representing the superior numbers indicating columns, to be read in vertical.
-pub(crate) fn get_column_numbers(height: usize, width: usize) -> String {
+pub(crate) fn get_column_numbers(height: usize, width: usize, use_emojis: bool) -> String {
     let (max_height_digits, max_width_digits) = ((height - 1).to_string().len(), (width - 1).to_string().len());
     // The space to leave on the left considering that will be occupied by row numbers below.
     let left_space = max_height_digits + ROW_NUMBER_RIGHT_SEPARATOR.len();
@@ -39,9 +39,17 @@ pub(crate) fn get_column_numbers(height: usize, width: usize) -> String {
     let mut result = String::with_capacity(max_width_digits * (left_space + width + 1) + 1);
     let mut i = 10_usize.pow((max_width_digits - 1) as u32);
     while i >= 1 {
-        result.push_str(&format!("{:\u{1F7EB}>width$}{}", "", ROW_NUMBER_RIGHT_SEPARATOR, width = max_height_digits));
+        if use_emojis {
+            result.push_str(&format!("{}{}", "🟫".repeat(max_height_digits), ROW_NUMBER_RIGHT_SEPARATOR));
+        } else {
+            result.push_str(&format!("{}{}", " ".repeat(max_height_digits), ROW_NUMBER_RIGHT_SEPARATOR));
+        }
         for j in 0..width {
-            result.push_str(&if j >= i || j == 0 && i == 1 { NUMBERS[j / i % 10].to_string() } else { String::from("🟫") });
+            if use_emojis {
+                result.push_str(&if j >= i || j == 0 && i == 1 { NUMBERS[j / i % 10].to_string() } else { String::from("🟫") });
+            } else {
+                result.push_str(&if j >= i || j == 0 && i == 1 { (j / i % 10).to_string() } else { String::from(" ") });
+            }
         }
         result.push('\n');
         i /= 10;
@@ -51,13 +59,17 @@ pub(crate) fn get_column_numbers(height: usize, width: usize) -> String {
 }
 
 
-pub(crate) fn get_row_number(number: usize, width: usize) -> String {
+pub(crate) fn get_row_number(number: usize, width: usize, use_emojis: bool) -> String {
     let number = number.to_string();
     let digits = number.len();
     let mut result = String::with_capacity(width);
-    result.push_str(String::from("🟫").repeat(width - digits).as_str());
+    result.push_str(String::from(if use_emojis { "🟫" } else { " " }).repeat(width - digits).as_str());
     for c in number.chars() {
-        result.push_str(NUMBERS[c.to_digit(10).unwrap() as usize]);
+        if use_emojis {
+            result.push_str(NUMBERS[c.to_digit(10).unwrap() as usize]);
+        } else {
+            result.push(c);
+        }
     }
     result
 }