+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 Many major changes:
 - Added utility functions to get `height`, `width` and number of mines from a `MineSweeper` object.
 - Added support for `wasm`, except for the `rand` crate which is not compatible with `wasm`.
-- Added format option `{:#}` to print the board with numbers.
+- Added format option `{:#}` to print the board as emojis.
+- Added format option `{:.0}` to print the board with row and column numbers.
 - Fixed an error that swapped `height` and `width` in the constructor.
 
 ## 0.2.0